+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- reqwest-based http client (enabled by default)
+
 ## [0.4.0] - 2021-03-08
 
 ### Fixed